Remove check to disable warnings in production builds
As per the Electron documentation, from Electron 2.0 on, developers will
only see warnings and recommendations printed to the developer console
when the binary's name is Electron, indicating that a developer is
currently looking at the console.

When we create our release builds the binary name is not Electron so
this check is redundant.