Move download snapshot logic to main process to avoid CORS errors
When making axios calls from the renderer process it uses XHR requests
which are susceptible to CORS errors (which in this case we were getting
when attempting to fetch the snapshot download from GitHub). When moving
this logic to the Electron main process it makes HTTP requests which
aren't susceptible to these pre-flight check problems.