Move transactionRecords data from global store to component
The transactionRecords data was only being used inside the one
TransactionList component. If we keep it in the store each time we loop
to fetch the data again it seems the app memory just grows with out
being garbage collected. This is still the case for other components
that loop and interact with the store, but if we don't need a piece of
data to be in the global store we should move it to the component that
requires it to have better memory efficiency.