Update SplashScreen.vue
updated text on splashscreen. I've noticed we have a lot of questions on "Initializing daemon... this may take some time!" proposing to change the wording to "Initializing daemon... this should take 5 minutes or less" This will be better for a new user.

"Connecting to peers... this should take 5 minutes or less" This will be better for a new user.